backend/requirements.txt
flask
gunicorn
google-genai
python-dotenv
datetime